auth: extract sandbox-backend data shapes into sandbox-backend-types.ts
Move the backend's type/interface declarations and NO_PASSWORD_SENTINEL
out of the God-file verbatim: public request/record types consumers
reach through the barrel (SeedUser, MintSessionRequest, AuthUserRecord,
CreateUserRequest, UpdateUserRequest, ProviderUserInfo,
SignInIdentitySpec, MintedSession) plus the internal per-record shapes
(StoredUser, Registration, BeforeStateReg, Mutable). Bodies verbatim;
the four previously file-private internal types gain 'export' so the
core can import them. The barrel re-exports the public surface + the
sentinel, so './sandbox-backend.js' importers are unchanged.
